+++ web-config/src/lib/configs/axis-config.svelte
